refactor: split the BankTransactions shell and finish the domain
BankTransactions.tsx is split along seams that were already private functions
inside it:
- BankTransactions/ -- the public entry. Name frozen; src/index.tsx repointed.
  Now just props, the preload hooks, and the 13-deep provider stack.
- BankTransactionsRouter/ -- was the private BankTransactionsContent: the route
  switch plus LockBankAccountFilter, which only it renders.
- ResponsiveBankTransactionsView/ -- was the private BankTransactionsTableView:
  owns the width measurement, the Container, the header, and the variant switch.

BankTransactionsTableContent -> BankTransactionsViewVariant. It names which view
renders, and two of its three values are not tables. The enum is internal, used
only by the shell and the header.

ResponsiveComponent is deliberately NOT adopted. It renders its own div,
measures that div, and keeps the resolved variant private, but this view needs
the variant in three places outside the switch: Container transparentBg,
BankTransactionsHeader tableContentMode, and the rule-suggestion dialog variant.
Adopting it would inject a DOM node inside Container -- a consumer-visible change
with no test covering it -- and duplicate either the measurement or the
Container+Header markup. ResponsiveCategorizationRulesView already calls
useSizeClass alongside ResponsiveComponent for the same reason, which says the
primitive is under-powered rather than its callers being sloppy.

COMPONENT_NAME stays 'bank-transactions' in the responsive view, so Container
still emits the published Layer__bank-transactions class; bankTransactions.scss
moved with it as responsiveBankTransactionsView.scss, contents unchanged.

src/components/ now holds only blocks, features, forms, ui, and utility.
