--- src/components/BankTransactions/BankTransactions.tsx
@@ -1,281 +1,0 @@
-import { type ReactNode, useCallback, useContext, useEffect, useRef, useState } from 'react'
-
-import type { LayerError } from '@utils/api/errorHandler'
-import { type BankTransactionFilters } from '@utils/bankTransactions/shared'
-import { BREAKPOINTS } from '@utils/screenSizeBreakpoints'
-import { unsafeAssertUnreachable } from '@utils/switch/assertUnreachable'
-import { usePreloadCategories } from '@api/businesses/[business-id]/categories/get'
-import { usePreloadCustomAccounts } from '@api/businesses/[business-id]/custom-accounts/get'
-import { usePreloadCustomers } from '@api/businesses/[business-id]/customers/get'
-import { usePreloadTagDimensions } from '@api/businesses/[business-id]/tags/dimensions/get'
-import { usePreloadVendors } from '@api/businesses/[business-id]/vendors/get'
-import { useElementSize } from '@hooks/utils/size/useElementSize'
-import { useIsVisible } from '@hooks/utils/visibility/useIsVisible'
-import { useBankAccountFilterActions } from '@providers/BankAccountsFilterStore/BankAccountsFilterStoreProvider'
-import { BankTransactionsCategorizationStoreProvider } from '@providers/BankTransactionsCategorizationStore/BankTransactionsCategorizationStoreProvider'
-import { BankTransactionsFeatureVisibilityProvider } from '@providers/BankTransactionsFeatureVisibility/BankTransactionsFeatureVisibilityProvider'
-import { BankTransactionsProvider } from '@providers/BankTransactionsPaginationProvider/BankTransactionsProvider'
-import { BankTransactionsPaginationProvider } from '@providers/BankTransactionsProvider/BankTransactionsPaginationProvider'
-import { BankTransactionsRoute, BankTransactionsRouteStoreProvider, useBankTransactionsRouteState } from '@providers/BankTransactionsRouteStore/BankTransactionsRouteStoreProvider'
-import { BulkSelectionStoreProvider, useCountSelectedIds } from '@providers/BulkSelectionStore/BulkSelectionStoreProvider'
-import { type BankTransactionsMode, LegacyModeProvider } from '@providers/LegacyModeProvider/LegacyModeProvider'
-import { useBankAccountsContext } from '@contexts/BankAccountsContext/BankAccountsContext'
-import {
-  useBankTransactionsContext,
-} from '@contexts/BankTransactionsContext/BankTransactionsContext'
-import { useBankTransactionsFiltersContext } from '@contexts/BankTransactionsFiltersContext/BankTransactionsFiltersContext'
-import { BankTransactionsFiltersProvider } from '@contexts/BankTransactionsFiltersContext/BankTransactionsFiltersProvider'
-import { BankTransactionsIsCategorizationEnabledProvider } from '@contexts/BankTransactionsIsCategorizationEnabledContext/BankTransactionsIsCategorizationEnabledContext'
-import { BankTransactionsStringOverridesProvider } from '@contexts/BankTransactionsStringOverridesContext/BankTransactionsStringOverridesContext'
-import { CategorizationRulesContext, CategorizationRulesProvider } from '@contexts/CategorizationRulesContext/CategorizationRulesContext'
-import { InAppLinkProvider, type LinkingMetadata } from '@contexts/InAppLinkContext'
-import { Container } from '@blocks/Layout/Container/Container'
-import { ErrorBoundary } from '@components/utility/ErrorBoundary/ErrorBoundary'
-import { BankTransactionsHeader } from '@features/bankTransactions/BankTransactionsHeader/BankTransactionsHeader'
-import { BankTransactionsList } from '@features/bankTransactions/BankTransactionsList/BankTransactionsList'
-import { BankTransactionsMobileList } from '@features/bankTransactions/BankTransactionsMobileList/BankTransactionsMobileList'
-import { BankTransactionsTable } from '@features/bankTransactions/BankTransactionsTable/BankTransactionsTable'
-import { BankTransactionsTableContent, type MobileComponentType } from '@features/bankTransactions/constants'
-import { type BankTransactionsStringOverrides } from '@features/bankTransactions/types'
-import { ResponsiveCategorizationRulesView } from '@features/categorization/ResponsiveCategorizationRulesView/ResponsiveCategorizationRulesView'
-import { SuggestedCategorizationRuleUpdatesDialog } from '@features/categorization/SuggestedCategorizationRuleUpdatesDialog/SuggestedCategorizationRuleUpdatesDialog'
-
-import './bankTransactions.scss'
-
-const COMPONENT_NAME = 'bank-transactions'
-
-export interface BankTransactionsProps {
-  asWidget?: boolean
-  pageSize?: number
-  /**
-   * @deprecated `mode` can be inferred from the bookkeeping configuration of a business
-   */
-  mode?: BankTransactionsMode
-
-  showCategorizationRules?: boolean
-  showCustomerVendor?: boolean
-  /**
-   * @deprecated This prop is no longer honored; transaction descriptions are always enabled.
-   */
-  showDescriptions?: boolean
-  /**
-   * @deprecated This prop is no longer honored; receipt uploads are always enabled.
-   */
-  showReceiptUploads?: boolean
-  showStatusToggle?: boolean
-  showTags?: boolean
-  showTooltips?: boolean
-  showUploadOptions?: boolean
-
-  applyGlobalDateRange?: boolean
-  monthlyView?: boolean
-
-  /**
-   * @deprecated `categorizeView` is no longer used. Categorization is enabled based on the bookkeeping configuration of a business.
-   */
-  categorizeView?: boolean
-  mobileComponent?: MobileComponentType
-  filters?: BankTransactionFilters
-  hideHeader?: boolean
-  collapseHeader?: boolean
-  stringOverrides?: BankTransactionsStringOverrides
-  renderInAppLink?: (details: LinkingMetadata) => ReactNode
-}
-
-export interface BankTransactionsWithErrorProps extends BankTransactionsProps {
-  onError?: (error: LayerError) => void
-}
-
-type BankTransactionsTableViewProps = Pick<
-  BankTransactionsProps,
-  'asWidget' | 'mobileComponent' | 'hideHeader' | 'collapseHeader'
->
-
-export const BankTransactions = ({
-  onError,
-  monthlyView = false,
-  applyGlobalDateRange = false,
-  mode,
-  renderInAppLink,
-  filters,
-  categorizeView: _categorizeView,
-  showDescriptions: _showDescriptions,
-  showReceiptUploads: _showReceiptUploads,
-
-  asWidget,
-  pageSize,
-  mobileComponent,
-  hideHeader,
-  collapseHeader,
-  stringOverrides,
-
-  ...featureVisibility
-}: BankTransactionsWithErrorProps) => {
-  usePreloadTagDimensions({ isEnabled: featureVisibility.showTags })
-  usePreloadCustomers({ isEnabled: featureVisibility.showCustomerVendor })
-  usePreloadVendors({ isEnabled: featureVisibility.showCustomerVendor })
-  usePreloadCustomAccounts({ isEnabled: featureVisibility.showUploadOptions })
-  usePreloadCategories()
-
-  return (
-    <ErrorBoundary onError={onError}>
-      <LegacyModeProvider overrideMode={mode}>
-        <BankTransactionsIsCategorizationEnabledProvider>
-          <CategorizationRulesProvider>
-            <BankTransactionsRouteStoreProvider>
-              <BankTransactionsFiltersProvider
-                monthlyView={monthlyView}
-                applyGlobalDateRange={applyGlobalDateRange}
-                filters={filters}
-              >
-                <BankTransactionsProvider>
-                  <BankTransactionsPaginationProvider pageSize={pageSize}>
-                    <BankTransactionsFeatureVisibilityProvider {...featureVisibility}>
-                      <BankTransactionsStringOverridesProvider stringOverrides={stringOverrides}>
-                        <InAppLinkProvider renderInAppLink={renderInAppLink}>
-                          <BulkSelectionStoreProvider>
-                            <BankTransactionsCategorizationStoreProvider>
-                              <BankTransactionsContent
-                                asWidget={asWidget}
-                                mobileComponent={mobileComponent}
-                                hideHeader={hideHeader}
-                                collapseHeader={collapseHeader}
-                              />
-                            </BankTransactionsCategorizationStoreProvider>
-                          </BulkSelectionStoreProvider>
-                        </InAppLinkProvider>
-                      </BankTransactionsStringOverridesProvider>
-                    </BankTransactionsFeatureVisibilityProvider>
-                  </BankTransactionsPaginationProvider>
-                </BankTransactionsProvider>
-              </BankTransactionsFiltersProvider>
-            </BankTransactionsRouteStoreProvider>
-          </CategorizationRulesProvider>
-        </BankTransactionsIsCategorizationEnabledProvider>
-      </LegacyModeProvider>
-    </ErrorBoundary>
-  )
-}
-
-const LockBankAccountFilter = () => {
-  const { setLocked } = useBankAccountFilterActions()
-
-  useEffect(() => {
-    setLocked(true)
-    return () => setLocked(false)
-  }, [setLocked])
-
-  return null
-}
-
-const BankTransactionsContent = (props: BankTransactionsTableViewProps) => {
-  const routeState = useBankTransactionsRouteState()
-  const { count } = useCountSelectedIds()
-
-  return (
-    <>
-      {count > 0 && <LockBankAccountFilter />}
-      {routeState.route === BankTransactionsRoute.BankTransactionsTable
-        ? <BankTransactionsTableView {...props} />
-        : <ResponsiveCategorizationRulesView />}
-    </>
-  )
-}
-
-type BankTransactionsTableContentViewProps = {
-  content: BankTransactionsTableContent
-}
-
-const BankTransactionsTableContentView = ({
-  content,
-}: BankTransactionsTableContentViewProps) => {
-  switch (content) {
-    case BankTransactionsTableContent.Table:
-      return <BankTransactionsTable />
-    case BankTransactionsTableContent.List:
-      return <BankTransactionsList />
-    case BankTransactionsTableContent.MobileList:
-      return <BankTransactionsMobileList />
-    default:
-      return unsafeAssertUnreachable({
-        value: content,
-        message: 'Unexpected table view content',
-      })
-  }
-}
-
-const BankTransactionsTableView = ({
-  asWidget = false,
-
-  mobileComponent,
-  hideHeader = false,
-  collapseHeader = false,
-}: BankTransactionsTableViewProps) => {
-  const scrollPaginationRef = useRef<HTMLDivElement>(null)
-  const isVisible = useIsVisible(scrollPaginationRef)
-
-  const { isMonthlyViewMode } = useBankTransactionsFiltersContext()
-
-  const { isLoading, hasMore, fetchMore } = useBankTransactionsContext()
-  const { isSyncing } = useBankAccountsContext()
-
-  const { setRuleSuggestion, ruleSuggestion } = useContext(CategorizationRulesContext)
-
-  useEffect(() => {
-    // Fetch more when the user scrolls to the bottom of the page
-    if (isMonthlyViewMode && isVisible && !isLoading && hasMore) {
-      fetchMore()
-    }
-  }, [isMonthlyViewMode, isVisible, isLoading, hasMore, fetchMore])
-
-  const handleRuleSuggestionOpenChange = useCallback((isOpen: boolean) => {
-    if (!isOpen) setRuleSuggestion(null)
-  }, [setRuleSuggestion])
-
-  const [listView, setListView] = useState(false)
-
-  const containerRef = useElementSize<HTMLDivElement>((size) => {
-    if (size.width > BREAKPOINTS.TABLET && listView) {
-      setListView(false)
-    }
-    else if (size.width <= BREAKPOINTS.TABLET && !listView) {
-      setListView(true)
-    }
-  })
-
-  const tableContentMode = listView && mobileComponent === 'mobileList'
-    ? BankTransactionsTableContent.MobileList
-    : listView
-      ? BankTransactionsTableContent.List
-      : BankTransactionsTableContent.Table
-
-  return (
-    <Container
-      className='Layer__Public'
-      transparentBg={listView && mobileComponent === 'mobileList'}
-      name={COMPONENT_NAME}
-      asWidget={asWidget}
-      ref={containerRef}
-    >
-      {!hideHeader && (
-        <BankTransactionsHeader
-          asWidget={asWidget}
-          tableContentMode={tableContentMode}
-          isSyncing={isSyncing}
-          collapseHeader={collapseHeader}
-        />
-      )}
-
-      <BankTransactionsTableContentView content={tableContentMode} />
-
-      <SuggestedCategorizationRuleUpdatesDialog
-        isOpen={!!ruleSuggestion}
-        onOpenChange={handleRuleSuggestionOpenChange}
-        ruleSuggestion={ruleSuggestion}
-        variant={tableContentMode === BankTransactionsTableContent.MobileList ? 'drawer' : 'modal'}
-      />
-
-      {isMonthlyViewMode ? <div ref={scrollPaginationRef} /> : null}
-    </Container>
-  )
-}
