+++ src/components/features/bankTransactions/ResponsiveBankTransactionsView/responsiveBankTransactionsView.scss
