+++ src/views/BankTransactionsWithLinkedAccounts/BankTransactionsWithLinkedAccounts.tsx
@@ -6,7 +6,7 @@ import { BankAccountsFilterStoreProvider, useSelectedBankAccountIds } from '@pro
 import { type BankTransactionsMode } from '@providers/LegacyModeProvider/LegacyModeProvider'
 import { type LinkingMetadata } from '@contexts/InAppLinkContext'
 import { View } from '@blocks/Layout/View/View'
-import { BankTransactions } from '@components/BankTransactions/BankTransactions'
+import { BankTransactions } from '@features/bankTransactions/BankTransactions/BankTransactions'
 import { type MobileComponentType } from '@features/bankTransactions/constants'
 import { type BankTransactionsStringOverrides } from '@features/bankTransactions/types'
 import { LinkedAccounts } from '@features/linkedAccounts/LinkedAccounts/LinkedAccounts'
